market(detail,candle)
Squashed commit of the following:

commit a8b74884d32469efea8bf32b0c1ff2ac9211499d
Author: peiliping <peilipingplp@gmail.com>
Date:   Fri Feb 21 12:56:10 2020 +0800

    market candle sink redis

commit 74bf0af10013ebe07a1f1e25d70708fd5deedc57
Author: peiliping <peilipingplp@gmail.com>
Date:   Fri Feb 21 12:46:27 2020 +0800

    market candle

commit 49fe7493505ab4ca4a2be0542144aa60a50fa5d8
Author: peiliping <peilipingplp@gmail.com>
Date:   Fri Feb 21 01:22:09 2020 +0800

    u

commit 920eb10fe78f207f49de72b7fbff48665133307c
Author: peiliping <peilipingplp@gmail.com>
Date:   Fri Feb 21 00:29:33 2020 +0800

    market candle

commit d77447629da3e68356e5485d6ba01bd33f60bb01
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 23:24:00 2020 +0800

    u

commit 0719516699c0686bcb3719683fd05014de864648
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 22:45:24 2020 +0800

    为滑窗 增加回调

commit c4ba52334c77a04bbb7adc52ab63f80d4cc31794
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 21:01:12 2020 +0800

    修复滑窗bug

commit f2edfab1d035bbf1cfb6f5576e8581186c1638e5
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 15:49:52 2020 +0800

    u

commit 1ad2bdb67eaa3d8e44e1126c75f289513c18e19c
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 11:59:38 2020 +0800

    u

commit 61b22ba1de592da8c9635b5898b0e1b9c5d66f4e
Author: peiliping <peilipingplp@gmail.com>
Date:   Thu Feb 20 11:41:24 2020 +0800

    market detail 支持redis的topic publish

commit dc13f3daca3a95503b3297491b56e8238c50b647
Author: peiliping <peilipingplp@gmail.com>
Date:   Wed Feb 19 23:45:54 2020 +0800

    market detail

commit c57f99de9935321ebc0f73898bd764c07bcb2fee
Author: peiliping <peilipingplp@gmail.com>
Date:   Wed Feb 19 18:58:45 2020 +0800

    修改离散策略

commit d1176323c472b8729cb8278f4914aa8b5e30dd05
Author: peiliping <peilipingplp@gmail.com>
Date:   Wed Feb 19 16:54:30 2020 +0800

    增加ordermatch的结果字段，表明驱动单的方向

+++ src/test/java/com/github/fevernova/task/exchangetestdata/Confrontation.java
@@ -40,13 +40,13 @@ public class Confrontation implements Runnable {
 
     final AtomicLong resultCount = new AtomicLong(0L);
 
-    final DataProvider<Long, OrderMatch> dataProvider = new DataProvider<Long, OrderMatch>() {
+    final DataProvider<Integer, OrderMatch> dataProvider = new DataProvider<Integer, OrderMatch>() {
 
 
         private OrderMatch orderMatch = (OrderMatch) new OrderMatchFactory().createData();
 
 
-        @Override public OrderMatch feedOne(Long key) {
+        @Override public OrderMatch feedOne(Integer key) {
 
             resultCount.incrementAndGet();
             return this.orderMatch;
